Fix pre_uninstall.py doesn't find deploy.conf
pre_uninstall.py is called by RPM uninstall (preun section).
It can't really find the deploy.conf. Lets teach pre_uninstall.py
where to find deploy.conf so that it can uninstall the RPM and
remove the generated virtualenv right.

Also remove --no-index from %post section of the RPM and let it
fetch from pypi to increase the probability of the RPM install
to be successful.