Generate rpms with the proper rpms to bootstrap the virtualenv even on newer releases of centos/rhel that have a python3 package and lack a python package.
Only use the virtualenv command if the interpreter is python 2.x which lacks the venv module.

Add a method to the plugins to apply updates to the configuration for the plugins, which is used to provide items to the template context.

Add more tests to the pipeline to verify packages work on different OS versions.